mentor/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-vertexai
langgraph
vertexai
google-auth
